refactor: extract split-layout core CSS to styles folder
