+++ packages/split-layout/src/styles/vaadin-split-layout-styles.d.ts
